+++ bootstrap/repository.php
@@ -11,10 +11,11 @@ namespace Ixocreate\Media;
 
 use Ixocreate\Database\Repository\RepositoryConfigurator;
 use Ixocreate\Media\Repository\MediaCreatedRepository;
+use Ixocreate\Media\Repository\MediaDefinitionInfoRepository;
 use Ixocreate\Media\Repository\MediaImageInfoRepository;
 use Ixocreate\Media\Repository\MediaRepository;
 
 /** @var RepositoryConfigurator $repository */
 $repository->addRepository(MediaRepository::class);
-$repository->addRepository(MediaImageInfoRepository::class);
+$repository->addRepository(MediaDefinitionInfoRepository::class);
 $repository->addRepository(MediaCreatedRepository::class);