+++ src/Handler/ImageHandler.php
@@ -9,14 +9,17 @@ declare(strict_types=1);
 
 namespace Ixocreate\Media\Handler;
 
-use Ixocreate\Filesystem\FilesystemManager;
+use Ixocreate\Filesystem\FilesystemInterface;
 use Ixocreate\Media\Config\MediaConfig;
+use Ixocreate\Media\Entity\MediaDefinitionInfo;
 use Ixocreate\Media\ImageDefinition\ImageDefinitionSubManager;
 use Ixocreate\Media\ImageDefinitionInterface;
 use Ixocreate\Media\MediaHandlerInterface;
 use Ixocreate\Media\MediaInterface;
 use Ixocreate\Media\MediaPaths;
 use Ixocreate\Media\Processor\ImageProcessor;
+use Ixocreate\Media\Repository\MediaDefinitionInfoRepository;
+use Ixocreate\Media\Repository\MediaRepository;
 
 final class ImageHandler implements MediaHandlerInterface
 {
@@ -41,9 +44,14 @@ final class ImageHandler implements MediaHandlerInterface
     ];
 
     /**
-     * @var ImageDefinitionSubManager
+     * @var MediaInterface
      */
-    private $imageDefinitionSubManager;
+    private $media;
+
+    /**
+     * @var string
+     */
+    private $mediaPath;
 
     /**
      * @var MediaConfig
@@ -51,25 +59,38 @@ final class ImageHandler implements MediaHandlerInterface
     private $mediaConfig;
 
     /**
-     * @var FilesystemManager
+     * @var MediaRepository
      */
-    private $filesystemManager;
+    private $mediaRepository;
+
+    /**
+     * @var MediaDefinitionInfoRepository
+     */
+    private $mediaDefinitionInfoRepository;
+
+    /**
+     * @var ImageDefinitionSubManager
+     */
+    private $imageDefinitionSubManager;
 
     /**
      * Image constructor.
      *
+     * @param MediaRepository $mediaRepository
+     * @param MediaDefinitionInfoRepository $mediaDefinitionInfoRepository
      * @param ImageDefinitionSubManager $imageDefinitionSubManager
      * @param MediaConfig $mediaConfig
-     * @param FilesystemManager $filesystemManager
      */
     public function __construct(
+        MediaRepository $mediaRepository,
+        MediaDefinitionInfoRepository $mediaDefinitionInfoRepository,
         ImageDefinitionSubManager $imageDefinitionSubManager,
-        MediaConfig $mediaConfig,
-        FilesystemManager $filesystemManager
+        MediaConfig $mediaConfig
     ) {
         $this->imageDefinitionSubManager = $imageDefinitionSubManager;
         $this->mediaConfig = $mediaConfig;
-        $this->filesystemManager = $filesystemManager;
+        $this->mediaDefinitionInfoRepository = $mediaDefinitionInfoRepository;
+        $this->mediaRepository = $mediaRepository;
     }
 
     /**
@@ -77,7 +98,7 @@ final class ImageHandler implements MediaHandlerInterface
      */
     public static function serviceName(): string
     {
-        return "Image";
+        return 'imageHandler';
     }
 
     /**
@@ -96,19 +117,46 @@ final class ImageHandler implements MediaHandlerInterface
     }
 
     /**
+     * * Directs a Image-File to the Image Processor (which transforms the Image with the given ImageDefinition Parameters)
+     * & creates + saves a MediaDefinitionInfo-Entity in the Database.
+     *
      * @param MediaInterface $media
-     * @throws \League\Flysystem\FileNotFoundException
+     * @param FilesystemInterface $filesystem
+     * @throws \Exception
      */
-    public function process(MediaInterface $media): void
+    public function process(MediaInterface $media, FilesystemInterface $filesystem): void
     {
-        $filesystem = $this->filesystemManager->get('media');
+        $this->media = $media;
+        $this->mediaPath = $this->media->publicStatus() ? MediaPaths::PUBLIC_PATH : MediaPaths::PRIVATE_PATH;
 
         foreach ($this->imageDefinitionSubManager->getServices() as $imageDefinitionClassName) {
             /** @var ImageDefinitionInterface $imageDefinition */
             $imageDefinition = $this->imageDefinitionSubManager->get($imageDefinitionClassName);
 
-            (new ImageProcessor($media, $imageDefinition, $this->mediaConfig, $filesystem))->process();
+            $process = (new ImageProcessor($this->media, $imageDefinition, $this->mediaConfig, $filesystem))->process();
+
+            if ($process === true) {
+
+                $file = $this->mediaPath . MediaPaths::IMAGE_DEFINITION_PATH . $imageDefinition->directory() . '/' . $this->media->basePath() . $this->media->filename();
+
+                $imageData = \getimagesizefromstring($filesystem->read($file));
+                $fileSize = $filesystem->getSize($file);
+
+                $mediaDefinitionInfo = new MediaDefinitionInfo([
+                    'mediaId' => $this->media->id(),
+                    'imageDefinition' => $imageDefinition::serviceName(),
+                    'width' => $imageData[0],
+                    'height' => $imageData[1],
+                    'fileSize' => $fileSize,
+                    'createdAt' => new \DateTimeImmutable(),
+                    'updatedAt' => new \DateTimeImmutable()
+                ]);
+
+                $this->mediaDefinitionInfoRepository->save($mediaDefinitionInfo);
+
+            }
         }
+        $this->updateMediaMetaData($filesystem);
     }
 
     /**
@@ -124,5 +172,29 @@ final class ImageHandler implements MediaHandlerInterface
             $directories[] = MediaPaths::IMAGE_DEFINITION_PATH . $imageDefinition->directory() . '/';
         }
         return $directories;
+    }
+
+    /**
+     * @param FilesystemInterface $filesystem
+     * @return void
+     */
+    private function updateMediaMetaData(FilesystemInterface $filesystem): void
+    {
+        $file = $this->mediaPath . $this->media->basePath() . $this->media->filename();
+
+        $imageData = \getimagesizefromstring($filesystem->read($file));
+
+        $metaData = [
+            'width' => $imageData[0],
+            'height' => $imageData[1]
+        ];
+
+        if ($this->media->metaData() === array()) {
+            $metaData = \array_merge($this->media->metaData(), $metaData);
+        }
+
+        $media = $this->media->with('metaData', $metaData);
+
+        $this->mediaRepository->save($media);
     }
 }